extract VerticleDeployer
